requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asttokens==2.4.1
attrs==23.1.0
backoff==2.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
clarifai==9.10.2
clarifai-grpc==9.10.0
click==8.1.7
cohere==4.35
colorama==0.4.6
comm==0.2.0
contextlib2==21.6.0
dataclasses-json==0.6.2
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
distro==1.8.0
executing==2.0.1
fastapi==0.104.1
fastavro==1.8.2
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.3
grpcio-tools==1.59.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.19.4
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.5.0
manifest-ml==0.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nlpcloud==1.1.45
nltk==3.8.1
numpy==1.26.2
omegaconf==2.2.3
opencv-python==4.7.0.68
openlm==0.0.5
packaging==23.2
pandas==2.1.3
parso==0.8.3
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.3.0
portalocker==2.8.2
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
pure-eval==0.2.2
pydantic==2.5.1
pydantic_core==2.14.3
Pygments==2.17.1
pytest==7.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-rapidjson==1.13
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
qdrant-client==1.6.9
redis==5.0.1
regex==2023.10.3
requests==2.31.0
rich==13.7.0
safetensors==0.4.0
schema==0.7.5
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlitedict==2.1.0
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.2
tritonclient==2.34.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.24.0.post1
wcwidth==0.2.11
yarl==1.9.3
zipp==3.17.0
